cabriot-admin-backend/admin-auth-service/requirements.txt
amqp==5.2.0
aniso8601==9.0.1
asgiref==3.6.0
async-timeout==4.0.3
billiard==4.2.0
celery==5.3.6
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.2
defusedxml==0.7.1
dj-database-url==2.1.0
Django==4.2
django-celery-results==2.5.1
django-cors-headers==3.14.0
django-dotenv==1.4.2
django-environ==0.10.0
django-multiselectfield==0.1.12
django-redis==5.4.0
django-rest-passwordreset==1.4.0
django-templated-mail==1.1.1
django-utils==0.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
graphene==3.3
graphene-django==3.2.0
graphql-core==3.2.3
graphql-relay==3.2.0
gunicorn==21.2.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.3.5
MarkupSafe==2.1.2
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==23.1
pika==1.3.2
pillow==10.2.0
promise==2.3
prompt-toolkit==3.0.43
psycopg2==2.9.9
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2023.3
redis==5.0.1
requests==2.28.2
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.8
simplejson==3.19.1
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
text-unidecode==1.3
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.6.0
